chore: move internal/highlight to cmd/frontend
Only frontend uses this package, so moving this package here to make clearer what service depends on this code.
This package also requires a call to the Init() func which only frontend does right now, so this also was a footgun where someone might start using this package outside of frontend, but then it's not correctly initialized.

Test plan: Just moved a package, go compiler doesn't complain.
