+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -249,6 +249,7 @@ go_library(
         "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/cody",
         "//cmd/frontend/internal/conf/validation",
+        "//cmd/frontend/internal/highlight",
         "//cmd/frontend/internal/inventory",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
@@ -297,7 +298,6 @@ go_library(
         "//internal/goroutine/recorder",
         "//internal/gosyntect",
         "//internal/gqlutil",
-        "//internal/highlight",
         "//internal/honey",
         "//internal/honey/search",
         "//internal/httpcli",
@@ -485,6 +485,7 @@ go_test(
         "//cmd/frontend/internal/auth/providers",
         "//cmd/frontend/internal/auth/userpasswd",
         "//cmd/frontend/internal/guardrails",
+        "//cmd/frontend/internal/highlight",
         "//cmd/frontend/internal/inventory",
         "//internal/actor",
         "//internal/api",
@@ -512,7 +513,6 @@ go_test(
         "//internal/gitserver/gitdomain",
         "//internal/gitserver/protocol",
         "//internal/gqlutil",
-        "//internal/highlight",
         "//internal/observation",
         "//internal/oobmigration",
         "//internal/ratelimit",
