+++ cmd/frontend/graphqlbackend/file.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend/externallink"
-	"github.com/sourcegraph/sourcegraph/internal/highlight"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/highlight"
 	"github.com/sourcegraph/sourcegraph/internal/ipynb"
 	"github.com/sourcegraph/sourcegraph/internal/markdown"
 )
