+++ cmd/frontend/internal/cli/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//cmd/frontend/internal/authz",
         "//cmd/frontend/internal/bg",
         "//cmd/frontend/internal/cli/middleware",
+        "//cmd/frontend/internal/highlight",
         "//cmd/frontend/internal/httpapi",
         "//cmd/frontend/oneclickexport",
         "//internal/actor",
@@ -62,7 +63,6 @@ go_library(
         "//internal/goroutine",
         "//internal/grpc",
         "//internal/grpc/defaults",
-        "//internal/highlight",
         "//internal/httpserver",
         "//internal/instrumentation",
         "//internal/jsonc",
