+++ cmd/frontend/internal/highlight/highlight_test.go
