+++ cmd/frontend/internal/highlight/mocks.go
