+++ cmd/frontend/internal/highlight/testdata/telemetry-raw.txt
