+++ cmd/frontend/internal/highlight/testdata/telemetry.scip
